cleaning up new contribution about MathHub
