+++ src/mmt-api/src/main/info/kwarc/mmt/api/archives/lmh/Git.scala
@@ -17,12 +17,8 @@ abstract class Git {
     * @param args Arguments to pass to git
     * @return
     */
-  def apply(wd: File, args: String*) : ShellCommand = {
-    val command = toArgs(args:_*)
-    // running shell commands is always brittle; better always print what we're trying to do 
-    println("running git command: " + command.mkString(" "))
-    println("in directory: " + wd.toString)
-    new ShellCommand(Some(wd), command, Map("GIT_TERMINAL_PROMPT" -> "0", "GIT_ASKPASS" -> "/bin/false"))({_ =>})({_ => })
+  def apply(wd: File, args: String*): ShellCommand.Result = {
+    ShellCommand.runInWithEnv(wd, Map("GIT_TERMINAL_PROMPT" -> "0", "GIT_ASKPASS" -> "/bin/false"), args:_*)
   }
 }
